requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain-openai
langchain
openai 